refactor: migrate commonTestResources from js to ts
