+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
   "eslint.workingDirectories": [
+    "commonTestResources",
     "packages/chai-openapi-response-validator",
     "packages/jest-openapi",
     "packages/openapi-validator"
