--- commonTestResources/config.js
@@ -1,5 +1,0 @@
-const port = 5000;
-
-module.exports = {
-  port,
-};
