+++ commonTestResources/tsconfig.eslint.json
@@ -1,0 +1,3 @@
+{
+  "include": ["**/*.ts"]
+}
