--- commonTestResources/utils.js
@@ -1,7 +1,0 @@
-const { inspect } = require('util');
-
-const joinWithNewLines = (...lines) => lines.join('\n\n');
-
-const str = (obj) => inspect(obj, { showHidden: false, depth: null });
-
-module.exports = { joinWithNewLines, str };
