+++ commonTestResources/utils.ts
@@ -1,0 +1,5 @@
+import { inspect } from 'util';
+
+export const joinWithNewLines = (...lines) => lines.join('\n\n');
+
+export const str = (obj) => inspect(obj, { showHidden: false, depth: null });
