+++ packages/chai-openapi-response-validator/test/assertions/satisfyApiSpec/differentRequestModules.test.ts
@@ -7,8 +7,7 @@ import supertest from 'supertest';
 import requestPromise from 'request-promise';
 
 import { str } from '../../../../../commonTestResources/utils';
-import app from '../../../../../commonTestResources/exampleApp';
-import { port } from '../../../../../commonTestResources/config';
+import app, { port } from '../../../../../commonTestResources/exampleApp';
 import chaiResponseValidator from '../../..';
 
 const appOrigin = `http://localhost:${port}`;
